compilers: Move the VisualStudioLikeCompiler class into mixins
