+++ mesonbuild/compilers/cpp.py
@@ -32,12 +32,12 @@ from .compilers import (
     PGICompiler,
     ArmclangCompiler,
     Compiler,
-    VisualStudioLikeCompiler,
 )
 from .c_function_attributes import CXX_FUNC_ATTRIBUTES, C_FUNC_ATTRIBUTES
 from .mixins.clike import CLikeCompiler
 from .mixins.ccrx import CcrxCompiler
 from .mixins.arm import ArmCompiler
+from .mixins.visualstudio import VisualStudioLikeCompiler
 
 
 def non_msvc_eh_options(eh, args):
